+++ src/clients/nycli/utils.h
@@ -69,7 +69,7 @@ gboolean playlist_exists (cli_infos_t *infos, gchar *playlist);
 void print_padding (gint length, gchar padchar);
 void print_indented (const gchar *string, guint level);
 gint find_terminal_width ();
-gchar *format_time (gint duration, gboolean use_hours);
+gchar *format_time (guint64 duration, gboolean use_hours);
 void enrich_mediainfo (xmmsv_t *val);
 
 #endif /* __UTILS_H__ */
